requirements.txt
gradio
faiss-cpu
numpy
cryptography
ollama
langchain_mistralai
